backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
supabase
sendgrid
twilio
apscheduler